Adds Dockerfile and docker-compose config
This is just a starting point for running SWT in docker as an alternative to installing
Python virtual environments.

Build image (if not already built) and start an interactive `bash` shell with the host's
current working directory mounted to `/usr/src/app/`. No ports forwarded by default, but
feel free to add if it makes sense. Note the `--rm` option will automatically remove the
image once it stops, which avoids clutter for unused images over time.

    docker-compose run --rm dev

Start it in one window, connect in another....

    # In first terminal...
    # Ctrl+C or closing termrinal window stops the container. Can view STDOUT.
    docker-compose up dev

    # In 2nd terminal...
    # Shell into running container.
    # Ctrl+C closes session, but does not stop container
    docker exec -it clams-kitchen-dev-1 bash

Rebuild the image as needed.

    docker-compose build --no-cache dev